requirements.txt
streamlit
langchain
groq
langchain-groq
langgraph
psycopg2-binary
python-dotenv
pandas
numpy
matplotlib
chromadb
langchain-community
langchainhub
faiss-cpu
sentence-transformers
# Content extraction dependencies
aiohttp
beautifulsoup4
pypdf
youtube_transcript_api
# Document loaders
langchain-text-splitters
pytesseract
# RAG components
langchain-core
langchain-openai
# PDF processing
pdfminer.six
# Core dependencies
validators
nest-asyncio